added ssh

+++ src/main/java/hospital/Lab/LabRequest.java
@@ -1,14 +1,14 @@
 package hospital.Lab;
 
 public class LabRequest {
-    private int record_id;
-    private int patient_id;
+    private long record_id;
+    private long patient_id;
     private String patient_name;
-    private int staff_id;
+    private long staff_id;
     private String doctor_name;
-    private int report_id;
+    private long report_id;
 
-    public LabRequest(int record_id, int patient_id, String patient_name, int staff_id, String doctor_name, int report_id) {
+    public LabRequest(long record_id, long patient_id, String patient_name, long staff_id, String doctor_name, long report_id) {
         this.record_id = record_id;
         this.patient_id = patient_id;
         this.patient_name = patient_name;
